Create pages to use React Router
