+++ src/App.js
@@ -1,25 +1,27 @@
 import React, { useEffect } from 'react';
-import { useDispatch, useSelector } from 'react-redux';
+import {
+  BrowserRouter as Router,
+  Switch,
+  Route,
+} from "react-router-dom";
+
 import { getDataForFeed } from './middlewares';
-import FeedContainer from './components/FeedContainer';
-import Filter from './components/Filter';
-import { changeFilter } from './actions';
+
+import Character from './pages/Character';
 
 function App() {
-  const feed = useSelector(state => state.feed);
-  const filter = useSelector(state => state.filter)
-  const currentCategory = useSelector(state => state.filter.currentCategory);
-  const dispatch = useDispatch();
-
-  useEffect(() => {
-    dispatch(getDataForFeed(currentCategory));
-  }, [dispatch]);
 
   return (
 
     <div>
-      <Filter changeFilter={changeFilter} dispatch={dispatch}/>
-      <FeedContainer feed={feed} filter={filter} />
+      <Router>
+        <Switch>
+          <Route path="/characters/:id">
+            <Character />
+          </Route>
+        </Switch>
+      </Router>
+
     </div>
   );
 }
